requirements.txt
python-dotenv
streamlit
beautifulsoup4
requests
python-docx
chardet
reportlab
langchain-text-splitters
langchain-groq
langchain-core
langchain-community
langchain-huggingface
numpy
-e .
